+++ app/src/main/java/org/linphone/utils/PermissionHelper.kt
@@ -69,4 +69,9 @@ class PermissionHelper private constructor(private val context: Context) {
     fun hasRecordAudioPermission(): Boolean {
         return hasPermission(Manifest.permission.RECORD_AUDIO)
     }
+
+    fun hasTelecomManagerPermissions(): Boolean {
+        return hasPermission(Manifest.permission.READ_PHONE_STATE) &&
+            hasPermission(Manifest.permission.MANAGE_OWN_CALLS)
+    }
 }